Move InstrumentConfiguration OTel convenience to separate file
